Lay out and add visuals for beginning of blog setup.
Adding visual elements as I add sections. This commit includes
background elements, the shrinking header, and the start of getting
the correct content into the correct places. Working on setting up
a grid and CSS organization that will scale well and make it easy
to edit/adjust elements as I go along.