Add layout structure to page, add more illustrations
This commit adds layout structure, including a container box that
is 1000px wide so that content stays structured on larger pages.
It also includes a new entry for roasted bananas along with
illustrations. I started to add content for further implementation
of the design as well (such as custard), etc.