Load correct content in featured-tip container
This PR adds the correct content and adjusts the way other content
is loaded in their respective containers to make sure it always
has a relevant post to the given section. Typographic styling and
padding for links will be included in next PR. Also adds a few
more illustrations (pistachio, sour cream..) and the inline-svg
gem.